Finish migration of api/web/push_subscriptions controller->request spec
